+++ src/main/java/pl/pawel/service/CustomerServiceImpl.java
@@ -30,7 +30,7 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public Customer getCustomerById(int id) {
+    public Customer getCustomerById(Long id) {
         return customerDao.getCustomerById(id);
     }
 
@@ -40,7 +40,7 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public void deleteCustomer(int id) {
+    public void deleteCustomer(Long id) {
         customerDao.deleteCustomer(id);
     }
 }
